Moved all the debug plotting into a new class KalmanPatRecPlots, turned on by steering file
